Crawling/requirements.txt
playwright>=1.42.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
streamlit>=1.32.0
pandas>=2.2.0
plotly>=5.20.0
python-dotenv>=1.0.0
